Refactor OccupancyMatrixPlot responsibilities
